decouple taskNew into smaller modules
