+++ packages/engine-core/src/tasks/global/new/utils.ts
@@ -1,0 +1,3 @@
+export const checkInputValue = (value: string | boolean): boolean => {
+    return value && typeof value === 'string' && value !== '' ? true : false;
+};
